+++ awc/tests/test_client.rs
@@ -719,7 +719,6 @@ async fn test_client_cookie_handling() {
 async fn client_unread_response() {
     let addr = test::unused_addr();
     let lst = std::net::TcpListener::bind(addr).unwrap();
-
     std::thread::spawn(move || {
         for stream in lst.incoming() {
             let mut stream = stream.unwrap();