+++ src/responder.rs
@@ -260,7 +260,7 @@ pub(crate) mod tests {
 
     #[actix_rt::test]
     async fn test_option_responder() {
-        let mut srv = init_service(
+        let srv = init_service(
             App::new()
                 .service(
                     web::resource("/none").to(|| async { Option::<&'static str>::None }),